auto_face_tracker/requirements.txt
opencv-python>=4.8.0
numpy>=1.21.0
pyserial>=3.5

